file reorganizing, controller support improvements
-Most of the menus are now navigable by gamepad and keyboard alone
-player movement/strafing is now analog
-renamed the prefabs folder to "objects" because the name "prefabs" was a holdover from my days as a Unity dev and does not make sense in a Godot environment
-a consistent naming scheme is now enforced:
Scripts follow C# naming conventions as usual, Scene files have capitalized first letters, resources are in snake_case, assets are in camelCase, and periods are used instead of underscores when categorizing things, as that allows you to double-click on a word in the file seperated by periods and select just the one word instead of the whole thing, very useful for renaming duplicates of files.

+++ Scripts/UI/DiffSlider.cs
@@ -18,9 +18,9 @@ public partial class DiffSlider : HSlider
 //  {
 //      
 //  }
-    private void OnDifficultyChange(int delta)
+    private void OnDifficultyChange(float delta)
     {
-        switch (delta)
+        switch ((int)delta)
         {
             case (-3):
             diffSetting.Text = "Game Journalist";
@@ -44,6 +44,6 @@ public partial class DiffSlider : HSlider
             diffSetting.Text = "Masochist";
             break;
         }
-        config.difficulty = delta;
+        config.difficulty = (int)delta;
     }
 }
